Add color support to OBJ export via MTL materials
When a model has painted color regions, OBJ export now generates an MTL
material file alongside the OBJ and bundles them in a ZIP download. Each
unique face color becomes a named material with Kd/Ka/Ks properties.
Unpainted faces get the default blue material. Without colors, export
remains a plain .obj file as before.

Also extracts the ZIP builder from threemf.ts into a shared zip.ts
module, and fixes main.ts to pass color data to OBJ export (was missing,
unlike the 3MF path which already did this).
