+++ src/export/obj.ts
@@ -1,10 +1,31 @@
 import type { MeshData } from '../geometry/types';
 import { downloadBlob, getExportFilename, getExportTitle } from './download';
+import { buildZip } from './zip';
+
+const DEFAULT_COLOR = '#4a9eff';
 
 export function exportOBJ(meshData: MeshData, customName?: string): void {
-  const { vertProperties, triVerts, numVert, numTri, numProp } = meshData;
+  const { vertProperties, triVerts, numVert, numTri, numProp, triColors } = meshData;
+  const title = getExportTitle();
 
-  const lines: string[] = [`# ${getExportTitle()}`];
+  // Detect whether we have painted triangles
+  let hasColors = false;
+  if (triColors) {
+    const painted = (triColors as Uint8Array & { _painted?: Uint8Array })._painted;
+    for (let t = 0; t < numTri; t++) {
+      const isPainted = painted
+        ? painted[t] === 1
+        : (triColors[t * 3] !== 0 || triColors[t * 3 + 1] !== 0 || triColors[t * 3 + 2] !== 0);
+      if (isPainted) { hasColors = true; break; }
+    }
+  }
+
+  const baseName = getExportFilename('obj', customName).replace(/\.obj$/, '');
+  const lines: string[] = [`# ${title}`];
+
+  if (hasColors) {
+    lines.push(`mtllib ${baseName}.mtl`);
+  }
 
   // Vertices
   for (let i = 0; i < numVert; i++) {
@@ -14,14 +35,80 @@ export function exportOBJ(meshData: MeshData, customName?: string): void {
     lines.push(`v ${x} ${y} ${z}`);
   }
 
-  // Faces (OBJ indices are 1-based)
-  for (let t = 0; t < numTri; t++) {
-    const i0 = triVerts[t * 3] + 1;
-    const i1 = triVerts[t * 3 + 1] + 1;
-    const i2 = triVerts[t * 3 + 2] + 1;
-    lines.push(`f ${i0} ${i1} ${i2}`);
-  }
+  if (hasColors && triColors) {
+    const painted = (triColors as Uint8Array & { _painted?: Uint8Array })._painted;
 
-  const blob = new Blob([lines.join('\n')], { type: 'text/plain' });
-  downloadBlob(blob, getExportFilename('obj', customName));
+    // Group triangles by color hex
+    const colorGroups = new Map<string, number[]>();
+
+    for (let t = 0; t < numTri; t++) {
+      const isPainted = painted
+        ? painted[t] === 1
+        : (triColors[t * 3] !== 0 || triColors[t * 3 + 1] !== 0 || triColors[t * 3 + 2] !== 0);
+
+      let hex: string;
+      if (isPainted) {
+        const r = triColors[t * 3].toString(16).padStart(2, '0');
+        const g = triColors[t * 3 + 1].toString(16).padStart(2, '0');
+        const b = triColors[t * 3 + 2].toString(16).padStart(2, '0');
+        hex = `#${r}${g}${b}`;
+      } else {
+        hex = DEFAULT_COLOR;
+      }
+
+      if (!colorGroups.has(hex)) colorGroups.set(hex, []);
+      colorGroups.get(hex)!.push(t);
+    }
+
+    // Write faces grouped by material (OBJ indices are 1-based)
+    for (const [hex, tris] of colorGroups) {
+      lines.push(`usemtl ${matName(hex)}`);
+      for (const t of tris) {
+        const i0 = triVerts[t * 3] + 1;
+        const i1 = triVerts[t * 3 + 1] + 1;
+        const i2 = triVerts[t * 3 + 2] + 1;
+        lines.push(`f ${i0} ${i1} ${i2}`);
+      }
+    }
+
+    // Generate MTL file
+    const mtlLines: string[] = [`# ${title} — Materials`];
+    for (const hex of colorGroups.keys()) {
+      const r = parseInt(hex.slice(1, 3), 16) / 255;
+      const g = parseInt(hex.slice(3, 5), 16) / 255;
+      const b = parseInt(hex.slice(5, 7), 16) / 255;
+      mtlLines.push(`newmtl ${matName(hex)}`);
+      mtlLines.push(`Kd ${r.toFixed(6)} ${g.toFixed(6)} ${b.toFixed(6)}`);
+      mtlLines.push(`Ka 0.100000 0.100000 0.100000`);
+      mtlLines.push(`Ks 0.300000 0.300000 0.300000`);
+      mtlLines.push(`Ns 40.000000`);
+      mtlLines.push(`d 1.000000`);
+      mtlLines.push('');
+    }
+
+    // Bundle OBJ + MTL in a ZIP for single-file download
+    const enc = new TextEncoder();
+    const zip = buildZip([
+      { name: `${baseName}.obj`, data: enc.encode(lines.join('\n')) },
+      { name: `${baseName}.mtl`, data: enc.encode(mtlLines.join('\n')) },
+    ]);
+
+    const blob = new Blob([zip], { type: 'application/zip' });
+    downloadBlob(blob, `${baseName}.zip`);
+  } else {
+    // No colors — plain OBJ
+    for (let t = 0; t < numTri; t++) {
+      const i0 = triVerts[t * 3] + 1;
+      const i1 = triVerts[t * 3 + 1] + 1;
+      const i2 = triVerts[t * 3 + 2] + 1;
+      lines.push(`f ${i0} ${i1} ${i2}`);
+    }
+
+    const blob = new Blob([lines.join('\n')], { type: 'text/plain' });
+    downloadBlob(blob, getExportFilename('obj', customName));
+  }
+}
+
+function matName(hex: string): string {
+  return hex === DEFAULT_COLOR ? 'Default' : `Color_${hex.slice(1)}`;
 }
